fix: 코드 리팩토링 및 리드미 수정
