+++ src/app/_hooks/useFashion/useCreate.ts
@@ -2,14 +2,20 @@ import { useRouter } from "next/navigation";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { toast } from "react-toastify";
 
-import { PostData } from "@/app/_types/type";
+import { inputType, PostData } from "@/app/_types/type";
 import { setFashionRoute } from "@/app/_utils/setFashionRoute";
 import { TAG_NAME } from "@/app/_constant/constant";
 import { createFashionApi } from "@/app/_api/fashionApi";
+import { useLoading } from "../useLoading";
+import { useUser } from "../useAuth";
+import imgCompression from "@/app/_utils/imgCompression";
 
 export default function useCreate() {
   const router = useRouter();
   const queryClient = useQueryClient();
+  const { isLoading: submitLoading, setLoading: setSubmitLoading } =
+    useLoading();
+  const { user } = useUser();
 
   const { mutate: createFashion } = useMutation({
     mutationFn: ({ user, title, content, tag, image }: PostData) =>
@@ -24,5 +30,33 @@ export default function useCreate() {
     },
   });
 
-  return { createFashion };
+  async function onSubmit(value: inputType) {
+    const { title, content, tag, imageFile } = value;
+
+    setSubmitLoading(true);
+
+    if (imageFile === null) return;
+
+    const compressionImage = await imgCompression(imageFile[0]);
+
+    if (!user) {
+      router.refresh();
+      return;
+    }
+
+    const fashionItemData = {
+      user,
+      title,
+      content,
+      tag,
+      image: compressionImage,
+    };
+    try {
+      createFashion(fashionItemData);
+    } catch {
+      setSubmitLoading(false);
+    }
+  }
+
+  return { onSubmit, submitLoading };
 }
